+++ tests/test_webhooks.py
@@ -24,6 +24,7 @@ from aiohttp.test_utils import AioHTTPTestCase, unittest_run_loop
 
 from crate.operator.webhooks import (
     WebhookClient,
+    WebhookClusterHealthPayload,
     WebhookEvent,
     WebhookInfoChangedPayload,
     WebhookPayload,
@@ -49,6 +50,7 @@ def test_payload_serialization_scale():
         upgrade_data=None,
         temporary_failure_data=None,
         info_data=None,
+        health_data=None,
     )
     assert json.loads(json.dumps(p)) == {
         "event": "scale",
@@ -64,6 +66,7 @@ def test_payload_serialization_scale():
         "upgrade_data": None,
         "temporary_failure_data": None,
         "info_data": None,
+        "health_data": None,
     }
 
 
@@ -79,6 +82,7 @@ def test_payload_serialization_upgrade():
         ),
         temporary_failure_data=None,
         info_data=None,
+        health_data=None,
     )
     assert json.loads(json.dumps(p)) == {
         "event": "upgrade",
@@ -94,6 +98,7 @@ def test_payload_serialization_upgrade():
         },
         "temporary_failure_data": None,
         "info_data": None,
+        "health_data": None,
     }
 
 
@@ -173,6 +178,7 @@ class TestWebhookClientSending(AioHTTPTestCase):
                 "upgrade_data": None,
                 "temporary_failure_data": None,
                 "info_data": None,
+                "health_data": None,
             },
         }
 
@@ -207,6 +213,7 @@ class TestWebhookClientSending(AioHTTPTestCase):
                 },
                 "temporary_failure_data": None,
                 "info_data": None,
+                "health_data": None,
             },
         }
 
@@ -236,6 +243,7 @@ class TestWebhookClientSending(AioHTTPTestCase):
                 "temporary_failure_data": {
                     "reason": "A snapshot is in progress",
                 },
+                "health_data": None,
             },
         }
 
@@ -263,6 +271,35 @@ class TestWebhookClientSending(AioHTTPTestCase):
                 "upgrade_data": None,
                 "info_data": {"external_ip": "192.168.1.10"},
                 "temporary_failure_data": None,
+                "health_data": None,
+            },
+        }
+
+    @unittest_run_loop
+    async def test_send_health_notification(self):
+        response = await self.webhook_client.send_notification(
+            "my-namespace",
+            "my-cluster",
+            WebhookEvent.HEALTH,
+            WebhookClusterHealthPayload(status="GREEN"),
+            WebhookStatus.SUCCESS,
+            logging.getLogger(__name__),
+        )
+        assert response.status == 200
+        data = await response.json()
+        assert data == {
+            "username": "itsme",
+            "password": "secr3t password",
+            "payload": {
+                "event": "health",
+                "status": "success",
+                "namespace": "my-namespace",
+                "cluster": "my-cluster",
+                "scale_data": None,
+                "upgrade_data": None,
+                "info_data": None,
+                "temporary_failure_data": None,
+                "health_data": {"status": "GREEN"},
             },
         }